reverted to Grid_ because of
Andrew, the idea was that by using Grid_host we were reasonably sure
job. I sort of think that 'Grid_' is about as short as we can go for
that part and then we need the gateway name.

Don't forget that sites already use that tag for other things....

I have to confess that length has already stuffed me up however, I only
reserved 20 char in the database and QUT chose not to use the recommendd
host names. Sigh ....

David

which is not satisfactory.  We'll have to wait for another chance to get this pinned down.